Create simpleTree.py
This creates a simple fractal tree with turtle. there s many ways in which we could control it with EEG parameters. line by line drawing is not so bad, as we can have them every second. Speed and color could depend on band power for instance. motor imagery could be used to prompt growth by increasing min_branch_length, and we could also de/fractalise it by puttign the angle under eeg control.